Various small changes
removed some buttons. Made it so that the user can't increase their battle stats increase their max stats.

+++ dbzdnd/dbzdnd/addValues.cs
@@ -31,11 +31,11 @@ namespace dbzdnd
             {
                 if(doing == "take")
                 {
-                    mainRetrunForm.changeCredits(0 - nudValue.Value);
+                    mainRetrunForm.changeCredits(0 - (int)nudValue.Value);
                 }
                 else if (doing == "add")
                 {
-                    mainRetrunForm.changeCredits(nudValue.Value);
+                    mainRetrunForm.changeCredits((int)nudValue.Value);
                 }
             }
             this.Close();
